goblinmini-docqa/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "goblin-docqa"
version = "0.1.0"
description = "Goblin Mini — local doc & code assistant to reduce Copilot usage"
authors = [{ name = "Fuaad", email = "you@example.com" }]
dependencies = [
  "fastapi>=0.95.0",
  "uvicorn[standard]>=0.22.0",
  "pydantic>=1.10.0",
  "requests>=2.31.0",
  "python-dotenv>=1.0.0",
  "jinja2>=3.1.0",
  "redis>=4.5.0",
  "aiohttp>=3.8.0",
  "python-multipart>=0.0.6",
  "prometheus-client>=0.19.0",
  "slowapi>=0.1.9",
  "psutil>=5.9.0",
]
[tool.setuptools.packages.find]
include = ["app*"]
[project.optional-dependencies]
testing = ["pytest", "httpx", "pytest-asyncio", "pytest-cov"]
local-model = ["torch>=2.0.0"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "C", "N", "Q"]
